Modifications to return JSON instead of a text array
Changed final result to collection
Changed insert to append (more readable and it's okay to throw the object right on in there)
Threw collection into our response object, called to_dict on it and used that as the return.
Modified the static_flag = True line to remove the '= True' for readability
Added hr. to the front of the imports
Added static_flag to the doc string