Change sent employee_id parameter to GET from a 'str' to a 'list'
When sending a string, once double digits came around, later on in the GET function, the code would attempt to loop through the list that it assumed was there. When this is a string instead of a list, it'd loop through the characters of the string instead meaning that there would be an error trying to delete 13 for example because it may attempt to delete 1 and 3 and either of which might not exist.