Implement Patch
Danger, not all of the things in this function have been tested. It's a lot of code, and there's a lot of places that it causes my eyes to spin. Everything looks okay, including the addition of new entries when needed and the change of entries instead of additions when needed too. I'm afraid I missed something or made a logic mistake somewhere, so keep an eye on this section of code! Potentially Dangerous! (I tested it as much as I could stomach manually... very late... very tired)